PythonClient/requirements.txt
msal==1.12.0
PyJWT==2.1.0
requests==2.23.0